Don't automatically restart the systemd service
This isn't the normal way that systemd service files are written, and
lots of admins will probably be surprised to find that this is
happening. Therefore, remove it from the default configuration.

This is especially relevant given that the default ExecStart command
produces a help message, which causes systemd to mark the unit as
failed. This causes an infinite restart loop on default installs.