Don't send stdout and stderr to custom locations
If the user is on systemd it's reasonable to assume that they are also
using journald. This is the default on most (all?) systemd systems and
any admin who configures their system to send things to syslog instead
of journald can deal with one more thing that needs to be
changed.

Therefore, optimize for the majority and send logs to the default
journald. Additionally, don't send stdout to /dev/null because this
isn't the norm with systemd; stdout can provide valuable debugging
information; and journald provides facilities for filtering out stdout
in log output.