Identify location lines based on _type field
This is related to issue #224. In the original file, it seemed that the search string "Z\t* " was not processed as regex by the strstr() function. Hence replacing the '*'-based regex pattern by a string to identify location data lines based on "_type". 

With this, the "bp" pointer had to start search the last '\t' in the data line from the beginning of the line.

I could not see from the strstr() source-code in GCC how it would process regex patterns at all ?