Make putrec hook also decide whether `last` and websocket receive a message
Current hook just governs whether the record is written, but this means that the `last` file
and the live websocket would still be updated with data that the user decided to filter. Now
we use the hook to skip those steps as well.
Update HOOKS.md to note the functionality change